+++ argopy/related/__init__.py
@@ -1,9 +1,13 @@
 from .topography import TopoFetcher
 from .ocean_ops_deployments import OceanOPSDeployments
+from .reference_tables import ArgoNVSReferenceTables
+from .argo_documentation import ArgoDocs
 
 #
 __all__ = (
     # Classes:
     "TopoFetcher",
     "OceanOPSDeployments",
+    "ArgoNVSReferenceTables",
+    "ArgoDocs",
 )
