+++ argopy/utilities.py
@@ -2705,243 +2705,6 @@ def get_ea_profile_page(WMO, CYC=None, **kwargs):
     return [url.format(this_id) for this_id in sorted(df["ID"])]
 
 
-class ArgoNVSReferenceTables:
-    """Argo Reference Tables
-
-    Utility function to retrieve Argo Reference Tables from a NVS server.
-
-    By default, this relies on: https://vocab.nerc.ac.uk/collection
-
-    Examples
-    --------
-    Methods:
-
-    >>> R = ArgoNVSReferenceTables()
-    >>> R.search('sensor')
-    >>> R.tbl(3)
-    >>> R.tbl('R09')
-
-    Properties:
-
-    >>> R.all_tbl_name
-    >>> R.all_tbl
-    >>> R.valid_ref
-
-    """
-    valid_ref = [
-        "R01",
-        "RR2",
-        "RD2",
-        "RP2",
-        "R03",
-        "R04",
-        "R05",
-        "R06",
-        "R07",
-        "R08",
-        "R09",
-        "R10",
-        "R11",
-        "R12",
-        "R13",
-        "R15",
-        "RMC",
-        "RTV",
-        "R16",
-        # "R18",
-        "R19",
-        "R20",
-        "R21",
-        "R22",
-        "R23",
-        "R24",
-        "R25",
-        "R26",
-        "R27",
-        # "R28",
-        # "R29",
-        # "R30",
-        "R40",
-    ]
-    """List of all available Reference Tables"""
-
-    def __init__(self,
-                 nvs="https://vocab.nerc.ac.uk/collection",
-                 cache: bool = True,
-                 cachedir: str = "",
-                 ):
-        """Argo Reference Tables from NVS"""
-        from .stores import httpstore
-        cachedir = OPTIONS["cachedir"] if cachedir == "" else cachedir
-        self.fs = httpstore(cache=cache, cachedir=cachedir)
-        self.nvs = nvs
-
-    def _valid_ref(self, rtid):
-        if rtid not in self.valid_ref:
-            rtid = "R%0.2d" % rtid
-            if rtid not in self.valid_ref:
-                raise ValueError(
-                    "Invalid Argo Reference Table, should be one in: %s"
-                    % ", ".join(self.valid_ref)
-                )
-        return rtid
-
-    def _jsConcept2df(self, data):
-        """Return all skos:Concept as class:`pandas.DataFrame`"""
-        content = {
-            "altLabel": [],
-            "prefLabel": [],
-            "definition": [],
-            "deprecated": [],
-            "id": [],
-        }
-        for k in data["@graph"]:
-            if k["@type"] == "skos:Collection":
-                Collection_name = k["alternative"]
-            elif k["@type"] == "skos:Concept":
-                content["altLabel"].append(k["altLabel"])
-                content["prefLabel"].append(k["prefLabel"]["@value"])
-                content["definition"].append(k["definition"]["@value"])
-                content["deprecated"].append(k["deprecated"])
-                content["id"].append(k["@id"])
-        df = pd.DataFrame.from_dict(content)
-        df.name = Collection_name
-        return df
-
-    def _jsCollection(self, data):
-        """Return last skos:Collection information as data"""
-        for k in data["@graph"]:
-            if k["@type"] == "skos:Collection":
-                name = k["alternative"]
-                desc = k["description"]
-                rtid = k["@id"]
-        return (name, desc, rtid)
-
-    def get_url(self, rtid, fmt="ld+json"):
-        """Return URL toward a given reference table for a given format
-
-        Parameters
-        ----------
-        rtid: {str, int}
-            Name or number of the reference table to retrieve. Eg: 'R01', 12
-        fmt: str, default: "ld+json"
-            Format of the NVS server response. Can be: "ld+json", "rdf+xml" or "text/turtle".
-
-        Returns
-        -------
-        str
-        """
-        rtid = self._valid_ref(rtid)
-        if fmt == "ld+json":
-            fmt_ext = "?_profile=nvs&_mediatype=application/ld+json"
-        elif fmt == "rdf+xml":
-            fmt_ext = "?_profile=nvs&_mediatype=application/rdf+xml"
-        elif fmt == "text/turtle":
-            fmt_ext = "?_profile=nvs&_mediatype=text/turtle"
-        else:
-            raise ValueError("Invalid format. Must be in: 'ld+json', 'rdf+xml' or 'text/turtle'.")
-        url = "{}/{}/current/{}".format
-        return url(self.nvs, rtid, fmt_ext)
-
-    @lru_cache
-    def tbl(self, rtid):
-        """Return an Argo Reference table
-
-        Parameters
-        ----------
-        rtid: {str, int}
-            Name or number of the reference table to retrieve. Eg: 'R01', 12
-
-        Returns
-        -------
-        class:`pandas.DataFrame`
-        """
-        rtid = self._valid_ref(rtid)
-        js = self.fs.open_json(self.get_url(rtid))
-        df = self._jsConcept2df(js)
-        return df
-
-    def tbl_name(self, rtid):
-        """Return name of an Argo Reference table
-
-        Parameters
-        ----------
-        rtid: {str, int}
-            Name or number of the reference table to retrieve. Eg: 'R01', 12
-
-        Returns
-        -------
-        tuple('short name', 'description', 'NVS id link')
-        """
-        rtid = self._valid_ref(rtid)
-        js = self.fs.open_json(self.get_url(rtid))
-        return self._jsCollection(js)
-
-    def search(self, txt, where='all'):
-        """Search for string in tables title and/or description
-
-        Parameters
-        ----------
-        txt: str
-        where: str, default='all'
-            Where to search, can be: 'title', 'description', 'all'
-
-        Returns
-        -------
-        list of table id matching the search
-        """
-        results = []
-        for tbl_id in self.all_tbl_name:
-            title = self.tbl_name(tbl_id)[0]
-            description = self.tbl_name(tbl_id)[1]
-            if where == 'title':
-                if txt.lower() in title.lower():
-                    results.append(tbl_id)
-            elif where == 'description':
-                if txt.lower() in description.lower():
-                    results.append(tbl_id)
-            elif where == 'all':
-                if txt.lower() in description.lower() or txt.lower() in title.lower():
-                    results.append(tbl_id)
-        return results
-
-    @property
-    def all_tbl(self):
-        """Return all Argo Reference tables
-
-        Returns
-        -------
-        OrderedDict
-            Dictionary with all table short names as key and table content as class:`pandas.DataFrame`
-        """
-        URLs = [self.get_url(rtid) for rtid in self.valid_ref]
-        df_list = self.fs.open_mfjson(URLs, preprocess=self._jsConcept2df)
-        all_tables = {}
-        [all_tables.update({t.name: t}) for t in df_list]
-        all_tables = collections.OrderedDict(sorted(all_tables.items()))
-        return all_tables
-
-    @property
-    def all_tbl_name(self):
-        """Return names of all Argo Reference tables
-
-        Returns
-        -------
-        OrderedDict
-            Dictionary with all table short names as key and table names as tuple('short name', 'description', 'NVS id link')
-        """
-        URLs = [self.get_url(rtid) for rtid in self.valid_ref]
-        name_list = self.fs.open_mfjson(URLs, preprocess=self._jsCollection)
-        all_tables = {}
-        [
-            all_tables.update({rtid.split("/")[-3]: (name, desc, rtid)})
-            for name, desc, rtid in name_list
-        ]
-        all_tables = collections.OrderedDict(sorted(all_tables.items()))
-        return all_tables
-
-
-
 @deprecated
 def cast_types(ds):  # noqa: C901
     """ Make sure variables are of the appropriate types according to Argo
@@ -3312,352 +3075,6 @@ def log_argopy_callerstack(level='debug'):
                 log.debug(msg)
             elif level == "warning":
                 log.warning(msg)
-
-
-class ArgoDocs:
-    """ADMT documentation helper class
-
-    Examples
-    --------
-    >>> ArgoDocs().list
-    >>> ArgoDocs().search("CDOM")
-    >>> ArgoDocs().search("CDOM", where='abstract')
-
-    >>> ArgoDocs(35385)
-    >>> ArgoDocs(35385).ris
-    >>> ArgoDocs(35385).abstract
-    >>> ArgoDocs(35385).show()
-    >>> ArgoDocs(35385).open_pdf()
-    >>> ArgoDocs(35385).open_pdf(page=12)
-
-    """
-    _catalogue = [
-        {
-            "category": "Argo data formats",
-            "title": "Argo user's manual",
-            "doi": "10.13155/29825",
-            "id": 29825
-        },
-        {
-            "category": "Quality control",
-            "title": "Argo Quality Control Manual for CTD and Trajectory Data",
-            "doi": "10.13155/33951",
-            "id": 33951
-        },
-        {
-            "category": "Quality control",
-            "title": "Argo quality control manual for dissolved oxygen concentration",
-            "doi": "10.13155/46542",
-            "id": 46542
-        },
-        {
-            "category": "Quality control",
-            "title": "Argo quality control manual for biogeochemical data",
-            "doi": "10.13155/40879",
-            "id": 40879
-        },
-        {
-            "category": "Quality control",
-            "title": "BGC-Argo quality control manual for the Chlorophyll-A concentration",
-            "doi": "10.13155/35385",
-            "id": 35385
-        },
-        {
-            "category": "Quality control",
-            "title": "BGC-Argo quality control manual for nitrate concentration",
-            "doi": "10.13155/84370",
-            "id": 84370
-        },
-        {
-            "category": "Quality control",
-            "title": "Quality control for BGC-Argo radiometry",
-            "doi": "10.13155/62466",
-            "id": 62466
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Argo DAC profile cookbook",
-            "doi": "10.13155/41151",
-            "id": 41151
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Argo DAC trajectory cookbook",
-            "doi": "10.13155/29824",
-            "id": 29824
-        },
-        {
-            "category": "Cookbooks",
-            "title": "DMQC Cookbook for Core Argo parameters",
-            "doi": "10.13155/78994",
-            "id": 78994
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Processing Argo oxygen data at the DAC level",
-            "doi": "10.13155/39795",
-            "id": 39795
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Processing Bio-Argo particle backscattering at the DAC level",
-            "doi": "10.13155/39459",
-            "id": 39459
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Processing BGC-Argo chlorophyll-A concentration at the DAC level",
-            "doi": "10.13155/39468",
-            "id": 39468
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Processing Argo measurement timing information at the DAC level",
-            "doi": "10.13155/47998",
-            "id": 47998
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Processing BGC-Argo CDOM concentration at the DAC level",
-            "doi": "10.13155/54541",
-            "id": 54541
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Processing Bio-Argo nitrate concentration at the DAC Level",
-            "doi": "10.13155/46121",
-            "id": 46121
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Processing BGC-Argo Radiometric data at the DAC level",
-            "doi": "10.13155/51541",
-            "id": 51541
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Processing BGC-Argo pH data at the DAC level",
-            "doi": "10.13155/57195",
-            "id": 57195
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Description of the Argo GDAC File Checks: Data Format and Consistency Checks",
-            "doi": "10.13155/46120",
-            "id": 46120
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Description of the Argo GDAC File Merge Process",
-            "doi": "10.13155/52154",
-            "id": 52154
-        },
-        {
-            "category": "Cookbooks",
-            "title": "BGC-Argo synthetic profile file processing and format on Coriolis GDAC",
-            "doi": "10.13155/55637",
-            "id": 55637
-        },
-        {
-            "category": "Cookbooks",
-            "title": "Argo GDAC cookbook",
-            "doi": "10.13155/46202",
-            "id": 46202
-        }
-    ]
-
-    class RIS:
-        """RIS file structure from TXT file"""
-
-        def __init__(self, file=None, fs=None):
-            self.record = None
-            self.fs = fs
-            if file:
-                self.parse(file)
-
-        def parse(self, file):
-            """Parse input file"""
-            # log.debug(file)
-
-            with self.fs.open(file, 'r', encoding="utf-8") as f:
-                TXTlines = f.readlines()
-            lines = []
-            # Eliminate blank lines
-            for line in TXTlines:
-                line = line.strip()
-                if len(line) > 0:
-                    lines.append(line)
-            TXTlines = lines
-
-            #
-            record = {}
-            for line in TXTlines:
-                # print("\n>", line)
-                if len(line) > 2:
-                    if line[2] == " ":
-                        tag = line[0:2]
-                        field = line[3:]
-                        # print("ok", {tag: field})
-                        record[tag] = [field]
-                    else:
-                        # print("-", line)
-                        record[tag].append(line)
-                elif len(line) == 2:
-                    record[line] = []
-                # else:
-                # print("*", line)
-
-            for key in record.keys():
-                record[key] = "; ".join(record[key])
-
-            self.record = record
-
-    @lru_cache
-    def __init__(self, docid=None, cache=False):
-        from .stores import httpstore
-
-        self.docid = None
-        self._ris = None
-        self._risfile = None
-        self._fs = httpstore(cache=cache, cachedir=OPTIONS['cachedir'])
-        self._doiserver = "https://dx.doi.org"
-        self._archimer = "https://archimer.ifremer.fr"
-
-        if isinstance(docid, int):
-            if docid in [doc['id'] for doc in self._catalogue]:
-                self.docid = docid
-            else:
-                raise ValueError("Unknown document id")
-        elif isinstance(docid, str):
-            start_with = lambda f, x: f[0:len(x)] == x if len(x) <= len(f) else False  # noqa: E731
-            if start_with(docid, '10.13155/') and docid in [doc['doi'] for doc in self._catalogue]:
-                self.docid = [doc['id'] for doc in self._catalogue if docid == doc['doi']][0]
-            else:
-                raise ValueError("'docid' must be an integer or a valid Argo DOI")
-
-    def __repr__(self):
-        summary = ["<argopy.ArgoDocs>"]
-        if self.docid is not None:
-            doc = [doc for doc in self._catalogue if doc['id'] == self.docid][0]
-            summary.append("Title: %s" % doc['title'])
-            summary.append("DOI: %s" % doc['doi'])
-            summary.append("url: https://dx.doi.org/%s" % doc['doi'])
-            summary.append("last pdf: %s" % self.pdf)
-            if 'AF' in self.ris:
-                summary.append("Authors: %s" % self.ris['AF'])
-            summary.append("Abstract: %s" % self.ris['AB'])
-        else:
-            summary.append("- %i documents with a DOI are available in the catalogue" % len(self._catalogue))
-            summary.append("- Use the method 'search' to find a document id")
-            summary.append("- Use the property 'list' to check out the catalogue")
-        return "\n".join(summary)
-
-    @property
-    def list(self):
-        """List of all available documents as a :class:`pandas.DataFrame`"""
-        return pd.DataFrame(self._catalogue)
-
-    @property
-    def js(self):
-        """Internal json record for a document"""
-        if self.docid is not None:
-            return [doc for doc in self._catalogue if doc['id'] == self.docid][0]
-        else:
-            raise ValueError("Select a document first !")
-
-    @property
-    def ris(self):
-        """RIS record of a document"""
-        if self.docid is not None:
-            if self._ris is None:
-                # Fetch RIS metadata for this document:
-                import re
-                file = self._fs.download_url("%s/%s" % (self._doiserver, self.js['doi']))
-                x = re.search(r'<a target="_blank" href="(https?:\/\/([^"]*))"\s+([^>]*)rel="nofollow">TXT<\/a>',
-                              str(file))
-                export_txt_url = x[1].replace("https://archimer.ifremer.fr", self._archimer)
-                self._risfile = export_txt_url
-                self._ris = self.RIS(export_txt_url, fs=self._fs).record
-            return self._ris
-        else:
-            raise ValueError("Select a document first !")
-
-    @property
-    def abstract(self):
-        """Abstract of a document"""
-        if self.docid is not None:
-            return self.ris['AB']
-        else:
-            raise ValueError("Select a document first !")
-
-    @property
-    def pdf(self):
-        """Link to the online pdf version of a document"""
-        if self.docid is not None:
-            return self.ris['UR']
-        else:
-            raise ValueError("Select a document first !")
-
-    def show(self, height=800):
-        """Insert document in pdf in a notebook cell
-
-        Parameters
-        ----------
-        height: int
-            Height in pixels of the cell
-        """
-        if self.docid is not None:
-            from IPython.core.display import HTML
-            return HTML(
-                '<embed src="%s" type="application/pdf" width="100%%" height="%ipx" />' % (self.ris['UR'], height))
-        else:
-            raise ValueError("Select a document first !")
-
-    def open_pdf(self, page=None, url_only=False):
-        """Open document in new browser tab
-
-        Parameters
-        ----------
-        page: int, optional
-            Open directly a specific page number
-        """
-        url = self.pdf
-        url += '#view=FitV&pagemode=thumbs'
-        if page:
-            url += '&page=%i' % page
-        if self.docid is not None:
-            if not url_only:
-                import webbrowser
-                webbrowser.open_new(url)
-            else:
-                return url
-        else:
-            raise ValueError("Select a document first !")
-
-    def search(self, txt, where='title'):
-        """Search for string in all documents title or abstract
-
-        Parameters
-        ----------
-        txt: str
-        where: str, default='title'
-            Where to search, can be 'title' or 'abstract'
-
-        Returns
-        -------
-        list
-
-        """
-        results = []
-        for doc in self.list.iterrows():
-            docid = doc[1]['id']
-            if where == 'title':
-                if txt.lower() in ArgoDocs(docid).js['title'].lower():
-                    results.append(docid)
-            elif where == 'abstract':
-                if txt.lower() in ArgoDocs(docid).abstract.lower():
-                    results.append(docid)
-        return results
 
 
 def drop_variables_not_in_all_datasets(ds_collection):
